requirements.txt
Django
django-tastypie
django-crispy-forms
django-widget-tweaks
